requirements.txt
phabricator==0.8.1
gitpython==3.1.20
click==8.0.1